Use a temporary Parse object to copy it
In order to correctly copy capabilities from the recursive region of
a Parse object, we must make sure that the copied part and a destination
buffer it is copied to have the same addresses modulo the capability
alignment.

Instead of appropriately aligning the destination buffer, let's use
a Parse object to make sure that any future changes in the Parse
structure would not introduce capability misalignments when copying the
recursive region.

Another solution to this problem would be to move the recursive region
to a separate structure but that would introduce additional changes
wherever Parse fields from the recursive regions are accessed.