Assume the correct Pager object alignment
Introduce the SQLITE_MALLOC_ALIGNMENT macro that indicates malloc
alignment which can be other than SQLITE_DEFAULT_ALIGNMENT, and use that
macro to assume that a pointer to a Pager object is appropriately
aligned under prior assertions that allow to make that assumption.
With that, a compiler can emit an inlined memcpy() call.

The pointer in question is returned by a malloc implementation that
aligns the returned pointer according to SQLITE_MALLOC_ALIGNMENT.
Later, the pointer is incremented by a number of bytes aligned to
SQLITE_MALLOC_ALIGNMENT.